Use Testcontainers for all Integration tests
