+++ .github/workflows/build.yaml
@@ -1,39 +1,9 @@
 name: Build
 on: [ push, pull_request ]
-env:
-  FLYWAY_VERSION: 7.15.0
-  FAF_DB_VERSION: v122
 jobs:
   test:
     runs-on: ubuntu-latest
     container: eclipse-temurin:17-jdk
-    services:
-      faf-db:
-        image: mariadb:10.6
-        ports:
-          - 3306:3306
-        options: >-
-          --health-cmd "mysqladmin ping --silent"
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-        env:
-          MYSQL_ROOT_PASSWORD: banana
-          MYSQL_DATABASE: faf
-          MYSQL_USER: faf-java-api
-          MYSQL_PASSWORD: banana
-      faf-league-db:
-        image: mariadb:10.6
-        options: >-
-          --health-cmd "mysqladmin ping --silent"
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-        env:
-          MYSQL_ROOT_PASSWORD: banana
-          MYSQL_DATABASE: faf-league
-          MYSQL_USER: faf-league-service
-          MYSQL_PASSWORD: banana
     steps:
       - name: Get the version
         id: get_version
@@ -47,22 +17,10 @@ jobs:
         run: apt update && apt install -y git wget
       - name: Checkout code
         uses: actions/checkout@v2
-      - name: Run flyway db migrations
-        env:
-          FLYWAY_URL: jdbc:mysql://faf-db/faf?useSSL=false
-          FLYWAY_USER: root
-          FLYWAY_PASSWORD: banana
-          FLYWAY_LOCATIONS: filesystem:db/migrations
-        run: |-
-          git clone --depth 1 --branch $FAF_DB_VERSION https://github.com/FAForever/db
-          wget -qO- https://repo1.maven.org/maven2/org/flywaydb/flyway-commandline/${FLYWAY_VERSION}/flyway-commandline-${FLYWAY_VERSION}-linux-x64.tar.gz | tar xvz && ln -s `pwd`/flyway-${FLYWAY_VERSION}/flyway /usr/local/bin
-          flyway migrate
       - name: Gradle build
         env:
           API_VERSION: ${{ steps.get_version.outputs.API_VERSION }}
           GRADLE_CLI_OPTS: ""
-          DATABASE_ADDRESS: faf-db
-          LEAGUE_DATABASE_ADDRESS: faf-league-db
         run: "chmod +x ./gradlew && ./gradlew ${GRADLE_CLI_OPTS} -Pversion=${API_VERSION} build jacocoTestReport"
       - name: Build and push Docker images
         if: github.ref == 'refs/heads/develop' || startsWith(github.ref, 'refs/tags')
