+++ src/inttest/java/com/faforever/api/league/LeagueScoreJournalElideTest.java
@@ -1,5 +1,6 @@
 package com.faforever.api.league;
 
+import com.faforever.api.AbstractIntegrationTest;
 import com.faforever.api.config.LeagueDatasourceConfig;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
@@ -17,7 +18,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD, scripts = "classpath:sql/league/truncateTables.sql", config = @SqlConfig(dataSource = LeagueDatasourceConfig.LEAGUE_DATA_SOURCE, transactionManager = LeagueDatasourceConfig.LEAGUE_TRANSACTION_MANAGER))
 @Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD, scripts = "classpath:sql/league/prepLeagueData.sql", config = @SqlConfig(dataSource = LeagueDatasourceConfig.LEAGUE_DATA_SOURCE, transactionManager = LeagueDatasourceConfig.LEAGUE_TRANSACTION_MANAGER))
-public class LeagueScoreJournalElideTest extends LeagueAbstractIntegrationTest {
+public class LeagueScoreJournalElideTest extends AbstractIntegrationTest {
 
   @Test
   void anyOneCanReadAllLeagueScoreJournal() throws Exception {
